requirements.txt
asgiref==3.3.4
astroid==2.4.1
autopep8==1.5
beautifulsoup4==4.9.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.5
Django==3.2.4
django-debug-toolbar==3.2.1
django-nose==1.4.7
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
nose==1.3.7
numpy==1.20.1
pandas==1.2.3
protobuf==3.11.3
psycopg2==2.8.6
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.5.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.23.0
six==1.14.0
soupsieve==2.2
SQLAlchemy==1.4.6
sqlparse==0.4.1
toml==0.10.0
tqdm==4.48.2
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
